例文紐付けAPI周り　型大幅変更
