+++ quizzer-lib/src/english/example/associateExample/dto.ts
@@ -1,0 +1,6 @@
+//
+export interface SubmitAssociationExampleAPIRequestDto {
+  wordName: string
+  checkedIdList: number[]
+  isAssociation: boolean
+}
