+++ quizzer-lib/src/english/example/index.ts
@@ -1,2 +1,3 @@
 export * from './addExample'
 export * from './searchExample'
+export * from './associateExample'
